backend/requirements.txt
asgiref==3.9.1
Django==5.2.6
sqlparse==0.5.3
tzdata==2025.2
djangorestframework==3.14.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0
django-cors-headers==4.2.0
pysentimiento
torch
transformers
google-generativeai